requirements.txt
pandas~=1.2.4
torch==1.8.1
torchvision==0.9.1
numpy~=1.19.1
PyYAML==5.3.1

scikit-learn~=0.24.1
requests~=2.24.0
tqdm~=4.50.2
Pillow~=8.2.0
opencv-python~=4.5.1.48
plotly~=4.14.3